A few fixes for 32-bit compiles
- Change the HAVE_SSE to HAVE_SSE4_2 and only provide it on
  x86-64.
- Cast two values that otherwise cause the compiler to warn on 32-bit

+++ engines/binject.c
@@ -146,7 +146,7 @@ static int fio_binject_getevents(struct thread_data *td, unsigned int min,
 		for (i = 0; i < events; i++) {
 			struct b_user_cmd *buc = (struct b_user_cmd *) buf + i;
 
-			bd->events[ev_index] = (struct io_u *) buc->usr_ptr;
+			bd->events[ev_index] = (struct io_u *) (unsigned long) buc->usr_ptr;
 			ev_index++;
 		}
 	}
